Fix domain eltype in Poisson PDE example
Also, revert ylim setting in recipe
