+++ examples/ODE_increaseprec.jl
@@ -1,8 +1,8 @@
-using ApproxFun #src
+using ApproxFun
 u = setprecision(1000) do
     d = BigFloat(0)..BigFloat(1)
     D = Derivative(d)
     [ldirichlet(); D-1] \ [1; 0]
-end
+end;
 using Test #src
 @test u(1) ≈ exp(BigFloat(1)) #src
