+++ examples/PDE_Helmholtz.jl
@@ -1,3 +1,12 @@
+# ## Helmholtz Equation
+
+# We solve
+# ```math
+# (Δ + 100)u(x,y) = 0
+# ```
+# on the rectangle `-1..1 × -1..1`,
+# subject to the condition that ``u=1`` on the boundary of the domain.
+
 using ApproxFun
 using LinearAlgebra
