+++ examples/system_of_eqn.jl
@@ -1,4 +1,4 @@
-# # Systems of equations
+# # System of equations
 
 # Systems of equations can be handled by creating a matrix of operators and functionals.
 # For example, we can solve the system
@@ -13,17 +13,7 @@
 
 # using the following code:
 
-using ApproxFun
-using LinearAlgebra
-
-x = Fun();
-B = Evaluation(Chebyshev(),-1);
-A = [B      0;
-     B*𝒟    0;
-     0      B;
-     𝒟^2-I  2I;
-     I      𝒟+I];
-u,v = A \ [0;0;0;exp(x);cos(x)];
+include("System1.jl")
 
 import Plots
 Plots.plot(u, label="u", xlabel="x", legend=:topleft)
