+++ test/ReadmeTest.jl
@@ -52,21 +52,33 @@ const EXAMPLES_DIR = joinpath(dirname(dirname(pathof(ApproxFun))), "examples")
 
     @testset "NonlinearBVP" begin
         include(joinpath(EXAMPLES_DIR, "NonlinearBVP1.jl"))
+        include(joinpath(EXAMPLES_DIR, "NonlinearBVP2.jl"))
     end
 
     @testset "ODE" begin
         include(joinpath(EXAMPLES_DIR, "ODE_BVP.jl"))
         include(joinpath(EXAMPLES_DIR, "ODE_increaseprec.jl"))
     end
+    @testset "System of Equations" begin
+        include(joinpath(EXAMPLES_DIR, "System1.jl"))
+    end
 
     @testset "Periodic" begin
         include(joinpath(EXAMPLES_DIR, "Periodic1.jl"))
+    end
+    @testset "Sampling" begin
+        include(joinpath(EXAMPLES_DIR, "Sampling1.jl"))
+    end
+    @testset "Eigenvalue" begin
+        include(joinpath(EXAMPLES_DIR, "Eigenvalue_standard.jl"))
+        include(joinpath(EXAMPLES_DIR, "Eigenvalue_symmetric.jl"))
     end
 
     # this is broken on v1.6 (on BandedArrays < v0.16.16), so we only test on higher versions
     if VERSION >= v"1.7"
         @testset "PDE" begin
-            include(joinpath(EXAMPLES_DIR, "PDE1.jl"))
+            include(joinpath(EXAMPLES_DIR, "PDE_Poisson.jl"))
+            include(joinpath(EXAMPLES_DIR, "PDE_Helmholtz.jl"))
         end
     end
